+++ preprocess/jsonify_multidata.py
@@ -1,61 +1,34 @@
+"""This script preprocess the multilingual dataset from 
+https://github.com/LIAAD/KeywordExtractor-Datasets in desired format
+to train a keyphrase extraction model using 
+https://github.com/thunlp/BERT-KPE approach"""
+
 import json
 import os
-import pdb
-import string
-import spacy
-from nltk.stem import SnowballStemmer
+from prepro_utils import remove_fullstop, find_sequence, lower_list
 
-stemmer = SnowballStemmer("spanish")
-nlp = spacy.load("es_core_news_sm")
-
+# define input and output paths 
+# (download data https://github.com/LIAAD/KeywordExtractor-Datasets)
 passage_folder = "../data/dataset/wicc/docsutf8/"
 keyphrase_folder = "../data/dataset/wicc/keys/"
-dest_file = "./data.json"
+dest_file = "./multidata.json"
 file_list = os.listdir(passage_folder)
-
-num = 0
-miss_num = 0
-
-
-def lemmatizer(text):
-    doc = nlp(text[0])
-    return " ".join([word.lemma_ for word in doc])
-
-
-def remove_punctuation(text):
-    return text.translate(str.maketrans("", "", string.punctuation))
-
-
-def remove_fullstop(text):
-    return text.replace(".", "")
-
-
-def find_sequence(seq, _list):
-    seq_list = seq
-    all_occurrence = [
-        idx
-        for idx in [i for i, x in enumerate(_list) if x == seq_list[0]]
-        if seq_list == _list[idx : idx + len(seq_list)]
-    ]
-    return -1 if not all_occurrence else all_occurrence
-
-
-def lower_list(_list):
-    return [stemmer.stem(x.lower()) for x in _list]
 
 
 with open(dest_file, "a") as jsonfile:
     for file_name in file_list:
-
-        print(file_name)
         a_dict = {}
         with open(passage_folder + file_name, "r") as f:
             text = f.read()
+            # removing fullstops
             text = remove_fullstop(text)
-            # text = lemmatizer(text)
+            # converting unicode to ascii
             str_en = text.encode("ascii", "ignore")
             str_de = str_en.decode()
+            # adding a list of words to the dictionary
             a_dict["doc_words"] = str_de.split()
+
+        # adding a list of keyphrases to the dictionary
         base = os.path.splitext(file_name)[0]
         with open(keyphrase_folder + base + ".key", "r") as f:
             text = f.readlines()
@@ -64,36 +37,36 @@ with open(dest_file, "a") as jsonfile:
             for line in text:
                 line = line.strip("\n")
                 if line:
+                    # converting unicode to ascii
                     str_en = line.encode("ascii", "ignore")
                     str_de = str_en.decode()
                     keyphrase_list.append(str_de.split())
 
-        num += 1
+        # Find occurances of keyphrases in the passage
         pos_list = []
         for keyphrase in keyphrase_list:
             kp_pos_list = []
             ind_list = find_sequence(
                 lower_list(keyphrase), lower_list(a_dict["doc_words"])
             )
+
+            # Some keyphrases are not detected in the passage
             if ind_list == -1:
-                miss_num += 1
                 keyphrase_list.remove(keyphrase)
                 continue
-                # pdb.set_trace()
             else:
-                # print(keyphrase)
-                # print(a_dict["doc_words"])
                 for ind in ind_list:
                     start_pos = ind
                     end_pos = ind + len(keyphrase) - 1
                     kp_pos_list.append([start_pos, end_pos])
                 pos_list.append(kp_pos_list)
 
+        # adding the keyphrases and their positions to the dictionary
         if keyphrase_list != []:
             a_dict["keyphrases"] = keyphrase_list
             a_dict["start_end_pos"] = pos_list
-            a_dict["url"] = file_name  # This acts as identifier while evaluating
+            # url acts as identifier while evaluating
+            a_dict["url"] = file_name  
+            # writing the dictionary to the json file
             jsonString = json.dump(a_dict, jsonfile)
             jsonfile.write("\n")
-
-# print(num, miss_num)
